+++ src/components/features/Order/atoms/OrderChacheCheckbox.tsx
@@ -1,7 +1,9 @@
 import { Checkbox, forwardRef } from '@chakra-ui/react';
 
-export const OrderCacheCheckbox = forwardRef((props, ref) => (
-  <Checkbox colorScheme="yellow" size="lg" ref={ref} {...props}>
-    현금영수증 신청
-  </Checkbox>
-));
+export const OrderCacheCheckbox = forwardRef((props, ref) => {
+  return (
+    <Checkbox colorScheme="yellow" size="lg" ref={ref} {...props}>
+      현금영수증 신청
+    </Checkbox>
+  );
+});
